refactor(teardown): fix methods and imports
