+++ tests/dashboard/teams/teams-invitations.spec.ts
@@ -14,6 +14,7 @@ import {
   waitSecondMessage,
 } from 'helpers/gmail';
 import { createInviteEmail } from 'helpers/teams/invite-email';
+import { createTeamName } from 'helpers/teams/create-team-name';
 
 let teamPage: TeamPage;
 let loginPage: LoginPage;
@@ -30,7 +31,7 @@ mainTest.beforeEach(async ({ page }: { page: Page }) => {
 });
 
 mainTest(qase(1164, 'Open the form via Invitations tab'), async () => {
-  const team = random().concat('autotest');
+  const team = createTeamName();
 
   await teamPage.createTeam(team);
   await teamPage.isTeamSelected(team);
@@ -43,7 +44,7 @@ mainTest(qase(1164, 'Open the form via Invitations tab'), async () => {
 registerTest(
   qase(1165, 'Open the form via Team Hero ("Invite members" button)'),
   async () => {
-    const team = random().concat('autotest');
+    const team = createTeamName();
 
     await teamPage.createTeam(team);
     await teamPage.isTeamSelected(team);
@@ -54,7 +55,7 @@ registerTest(
 );
 
 mainTest(qase(1166, 'Invite via owner (single invitation, editor)'), async () => {
-  const team = random().concat('autotest');
+  const team = createTeamName();
   const email = createInviteEmail('editor');
 
   await teamPage.createTeam(team);
@@ -74,7 +75,7 @@ mainTest(qase(1166, 'Invite via owner (single invitation, editor)'), async () =>
 });
 
 mainTest(qase(1167, 'Invite via owner (single invitation, admin)'), async () => {
-  const team = random().concat('autotest');
+  const team = createTeamName();
   const email = createInviteEmail('admin');
 
   await teamPage.createTeam(team);
@@ -97,7 +98,7 @@ mainTest(qase(1167, 'Invite via owner (single invitation, admin)'), async () => 
 });
 
 mainTest(qase(1175, 'Fail to send invitation to existing team member'), async () => {
-  const team = random().concat('autotest');
+  const team = createTeamName();
 
   await teamPage.createTeam(team);
   await teamPage.isTeamSelected(team);
@@ -119,7 +120,7 @@ mainTest(qase(1175, 'Fail to send invitation to existing team member'), async ()
 registerTest.describe(
   'As Owner - Invite Members - Send invite (registered but not team member)',
   () => {
-    const team = random().concat('autotest');
+    const team = createTeamName();
 
     registerTest.beforeEach(
       'Create new account, logout, login with main account, create team and open invitations page',
@@ -177,7 +178,7 @@ registerTest.describe(
 mainTest.describe(
   'As Owner - Invite Members - Send multiple invites (Editor role)',
   () => {
-    const team = random().concat('autotest');
+    const team = createTeamName();
     const firstEditor = random().concat('autotest');
     const secondEditor = random().concat('autotest');
 
@@ -273,7 +274,7 @@ mainTest.describe(
 );
 
 mainTest(qase(1176, 'Resend multiple invitations via owner'), async () => {
-  const team = random().concat('autotest');
+  const team = createTeamName();
   const email1 = createInviteEmail('editor', 'editor1');
   const email2 = createInviteEmail('editor', 'editor2');
   const email3 = createInviteEmail('editor', 'editor3');
@@ -297,7 +298,7 @@ mainTest(qase(1176, 'Resend multiple invitations via owner'), async () => {
 });
 
 mainTest(qase(1178, 'Delete multiple invitations via owner'), async () => {
-  const team = random().concat('autotest');
+  const team = createTeamName();
   const email1 = createInviteEmail('editor', 'editor1');
   const email2 = createInviteEmail('editor', 'editor2');
   const email3 = createInviteEmail('editor', 'editor3');
@@ -321,7 +322,7 @@ mainTest(qase(1178, 'Delete multiple invitations via owner'), async () => {
 });
 
 mainTest(qase(1181, 'Change role in invitation via owner'), async () => {
-  const team = random().concat('autotest');
+  const team = createTeamName();
   const email = createInviteEmail('editor', 'role');
 
   await teamPage.createTeam(team);
